pyproject.toml
[tool.poetry]
name = "optibonds"
version = "0.1.0"
description = "Build, optimize, and visualize bond ladders."
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "optibonds"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
requests = "^2.31.0"
numpy-financial = "^1.0.0"
tabulate = "^0.9.0"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
